+++ kmptoolkit-platform/src/androidMain/kotlin/io/github/jamal_wia/kmptoolkit/platform/crash/Time.android.kt
@@ -1,3 +1,0 @@
-package io.github.jamal_wia.kmptoolkit.platform.crash
-
-internal actual fun currentTimeMillis(): Long = System.currentTimeMillis()